+++ config/settings.py
@@ -53,6 +53,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'django.forms',
+    'django.contrib.humanize',
 
     'sso',
     'core.apps.CoreConfig',